fix(review): verifier 포트 28090대 이동·소스 메시지 API 생성·limit ::integer 캐스트 (#597)
- 28040~43은 verify_attachment_upload와 충돌 — 28090~93으로 이동
- 시드에는 message 행이 없음 — verify_memory_plane 패턴으로 API POST 생성
- Swift Int는 bigint로 바인딩되어 memory_search_hybrid(integer) 해석 실패(500) — ::integer 캐스트

+++ server/Sources/MomoServer/Routes/MemoryRoutes.swift
@@ -159,7 +159,7 @@ struct MemoryRoutes: Sendable {
                   FROM memory_search_hybrid(
                          \(workspaceID), \(principal.memberID), \(query),
                          \(vectorLiteral)::vector(384), \(scope)::text,
-                         \(agentID)::uuid, \(limit), 60
+                         \(agentID)::uuid, \(limit)::integer, 60
                        ) search
                  ORDER BY search.rrf_score DESC, search.memory_id
                 """,
